# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  responses.CommonResponse:
    properties:
      data: {}
      errors: {}
      request_id:
        type: string
      response_code:
        type: integer
      response_message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive RESTful API built with Go, Fiber, and PostgreSQL using
    Hexagonal Architecture
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go RESTful API - Hexagonal Architecture
  version: "1.0"
paths:
...
  /media:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all media files with pagination
      parameters:
      - default: 10
        description: 'Number of media items to return (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 'Number of media items to skip (default: 0)'
        in: query
        minimum: 0
        name: offset
        type: integer
      - description: Search query to filter media by name or filename
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of media files
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
        "400":
          description: Bad request - Invalid parameters
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Get all media
      tags:
      - media
    post:
      consumes:
      - multipart/form-data
      description: Upload a new media file to the system
      parameters:
      - description: Media file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Custom name for the media file
        in: formData
        name: name
        type: string
      - description: Description of the media file
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Media uploaded successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad request - Invalid file or parameters
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "413":
          description: Payload too large - File size exceeds limit
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "415":
          description: Unsupported media type
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Upload media
      tags:
      - media
  /media/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a media file from the system
      parameters:
      - description: Media ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Media deleted successfully
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "400":
          description: Bad request - Invalid media ID
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "404":
          description: Not found - Media does not exist
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Delete media
      tags:
      - media
    get:
      consumes:
      - application/json
      description: Retrieve a specific media file by its unique identifier
      parameters:
      - description: Media ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Media file details
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad request - Invalid media ID
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "404":
          description: Not found - Media does not exist
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Get media by ID
      tags:
      - media
    put:
      consumes:
      - application/json
      description: Update media file metadata (name, description, etc.)
      parameters:
      - description: Media ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Media update request
        in: body
        name: media
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Media updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad request - Invalid input data
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "404":
          description: Not found - Media does not exist
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Update media
      tags:
      - media
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"